fix non-determinism in tests
