+++ packages/houdini/legacy/codegen/generators/typescript/documentTypes.ts
@@ -117,92 +117,6 @@ export async function generateDocumentTypes(config: Config, docs: Document[]) {
 			typePaths.push(typeDefPath)
 		})
 	)
-
-	// now that we have every type generated, create an index file in the runtime root that exports the types
-	const typeIndex = AST.program(
-		typePaths
-			.sort((a, b) => a.localeCompare(b))
-			.map((typePath) => {
-				return AST.exportAllDeclaration(
-					AST.literal(
-						'./' +
-							path
-								.relative(path.resolve(config.typeIndexPath, '..'), typePath)
-								// remove the .d.ts from the end of the path
-								.replace(/\.[^/.]+\.[^/.]+$/, '')
-					),
-					null
-				)
-			})
-			.concat([
-				AST.exportAllDeclaration(AST.literal('./runtime'), null),
-				AST.exportAllDeclaration(AST.literal('./graphql'), null),
-			])
-	)
-
-	// stringify the value so we can push it through the plugins
-	const exportDefaultAs = ({ module, as }: { module: string; as: string }) =>
-		`\nexport { default as ${as} } from "${module}"\n`
-	const exportStarFrom = ({ module }: { module: string }) => `\nexport * from "${module}"\n`
-	let { code: indexContent } = await printJS(typeIndex)
-	for (const plugin of config.plugins) {
-		if (plugin.indexFile) {
-			indexContent = plugin.indexFile({
-				config,
-				content: indexContent,
-				exportDefaultAs,
-				exportStarFrom,
-				pluginRoot: config.pluginDirectory(plugin.name),
-				typedef: true,
-				documents: docs,
-			})
-		}
-
-		// if the plugin generated a runtime
-		if (plugin.includeRuntime) {
-			indexContent += exportStarFrom({
-				module:
-					'./' +
-					path.relative(config.rootDir, config.pluginRuntimeDirectory(plugin.name)),
-			})
-		}
-	}
-
-	// write the contents
-	await fs.writeFile(config.typeIndexPath, indexContent)
-
-	// remove any runtime scalars from the list of scalars missing config
-	const missingScalarNames = [...missingScalars].filter(
-		(scalar) => !config.configFile.features?.runtimeScalars?.[scalar]
-	)
-
-	// if we were missing scalars, we need to warn the user and tell them
-	if (missingScalarNames.length > 0) {
-		console.warn(`⚠️  Missing definitions for the following scalars: ${[
-			...missingScalarNames,
-		].join(', ')}
-Generated types will contain an any type in place of these values. To fix this, provide an equivalent
-type in your config file:
-
-{
-  scalars: {
-    /* in your case, something like */
-${[...missingScalarNames]
-	.map(
-		(c) =>
-			`    ${c}: {                  // <- The GraphQL Scalar
-      type: "YourType_${c}"  // <-  The TypeScript type
-    }`
-	)
-	.join(
-		`,
-`
-	)}
-  }
-}
-
-For more information, please visit this link: ${siteURL}/api/config#custom-scalars`)
-	}
 }
 
 function convertToTs(source: ExpressionKind): TSTypeKind {
